Bug fixes & new wanna eat button behavior & linter style changes

+++ src/handlers/storage_list/storage_list_delete.cpp
@@ -3,8 +3,7 @@
 #include "backend/id_types.hpp"
 #include "handlers/common.hpp"
 #include "render/storage_list/storage_list_render.hpp"
-
-#include <sstream>
+#include "utils.hpp"
 
 namespace cookcookhnya::handlers::storage_delete {
 
@@ -15,7 +14,9 @@ void deleteStorage(
     bot.answerCallbackQuery(cq.id);
     if (cq.data.starts_with("st")) {
         auto storageId = utils::parseSafe<api::StorageId>(cq.data.substr(4));
-        storageApi.delete_(cq.from->id, *storageId);
+        if (storageId) {
+            storageApi.delete_(cq.from->id, *storageId);
+        }
     }
     renderStorageList(true, cq.from->id, cq.message->chat->id, bot, storageApi);
     stateManager.put(StorageList{});
